+++ src/elements/Voucher.php
@@ -495,7 +495,7 @@ class Voucher extends Purchasable
         if (empty($this->sku)) {
             try {
                 $voucherType = GiftVoucher::$plugin->getVoucherTypes()->getVoucherTypeById($this->typeId);
-                $this->sku = Craft::$app->getView()->renderObjectTemplate($voucherType->skuFormat, $this);
+                $this->sku = GiftVoucher::$plugin->getTemplates()->renderSandboxedObjectTemplate((string)$voucherType->skuFormat, $this);
             } catch (\Exception) {
                 $this->sku = '';
             }
@@ -575,7 +575,7 @@ class Voucher extends Purchasable
         $description = $this->title;
 
         if ($format = $this->getType()->descriptionFormat) {
-            if ($rendered = Craft::$app->getView()->renderObjectTemplate($format, $this)) {
+            if ($rendered = GiftVoucher::$plugin->getTemplates()->renderSandboxedObjectTemplate($format, $this)) {
                 $description = $rendered;
             }
         }